# frontend/src/shared/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SkillDraft": {
        "additionalProperties": true,
        "description": "A proposed skill package awaiting expert review; never active until promoted.",
        "properties": {
          "author": {
            "$ref": "#/components/schemas/SkillDraftAuthor"
          },
          "catalog_sha256": {
            "title": "Catalog Sha256",
            "type": "string"
          },
          "content_sha256": {
            "title": "Content Sha256",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "draft_id": {
            "title": "Draft Id",
            "type": "string"
          },
          "manifest": {
            "additionalProperties": true,
            "title": "Manifest",
            "type": "object"
          },
          "rationale": {
            "title": "Rationale",
            "type": "string"
          },
          "references": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "References"
          },
          "review": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SkillDraftReview"
              },
              {
                "type": "null"
              }
            ]
          },
          "skill_id": {
            "title": "Skill Id",
            "type": "string"
          },
          "skill_md": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skill Md"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "draft_id",
          "skill_id",
          "version",
          "status",
          "author",
          "rationale",
          "manifest",
          "content_sha256",
          "catalog_sha256",
          "review",
          "created_at",
          "updated_at"
        ],
        "title": "SkillDraft",
        "type": "object"
      },
      "SkillDraftAuthor": {
        "additionalProperties": true,
        "properties": {
          "actor": {
            "title": "Actor",
            "type": "string"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "subject_id": {
            "title": "Subject Id",
            "type": "string"
          }
        },
        "required": [
          "actor",
          "subject_id",
          "session_id",
          "run_id"
        ],
        "title": "SkillDraftAuthor",
        "type": "object"
      },
      "SkillDraftList": {
        "additionalProperties": false,
        "properties": {
          "drafts": {
            "items": {
              "$ref": "#/components/schemas/SkillDraft"
            },
            "title": "Drafts",
            "type": "array"
          }
        },
        "required": [
          "drafts"
        ],
        "title": "SkillDraftList",
        "type": "object"
      },
      "SkillDraftPackage": {
        "additionalProperties": false,
        "description": "An approved draft as files, the input to ``scripts/promote_skill_draft.py``.",
        "properties": {
          "draft_id": {
            "title": "Draft Id",
            "type": "string"
          },
          "files": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Files",
            "type": "object"
          },
          "package_sha256": {
            "title": "Package Sha256",
            "type": "string"
          },
          "review": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SkillDraftReview"
              },
              {
                "type": "null"
              }
            ]
          },
          "skill_id": {
            "title": "Skill Id",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "draft_id",
          "skill_id",
          "version",
          "review",
          "files",
          "package_sha256"
        ],
        "title": "SkillDraftPackage",
        "type": "object"
      },
      "SkillDraftRequest": {
        "additionalProperties": false,
        "description": "A skill package proposed for review (W9-11): the same three parts a\nshipped skill has, with the manifest's status held at ``draft``.",
        "properties": {
          "manifest": {
            "additionalProperties": true,
            "title": "Manifest",
            "type": "object"
          },
          "rationale": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Rationale",
            "type": "string"
          },
          "references": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "References",
            "type": "object"
          },
          "skill_md": {
            "maxLength": 20000,
            "minLength": 1,
            "title": "Skill Md",
            "type": "string"
          }
        },
        "required": [
          "skill_md",
          "manifest",
          "rationale"
        ],
        "title": "SkillDraftRequest",
        "type": "object"
      },
      "SkillDraftReview": {
        "additionalProperties": true,
        "properties": {
          "at": {
            "title": "At",
            "type": "string"
          },
          "decision": {
            "title": "Decision",
            "type": "string"
          },
          "note": {
            "title": "Note",
            "type": "string"
          },
          "reviewer": {
            "title": "Reviewer",
            "type": "string"
          }
        },
        "required": [
          "reviewer",
          "decision",
          "note",
          "at"
        ],
        "title": "SkillDraftReview",
        "type": "object"
      },
      "SkillDraftReviewRequest": {
        "additionalProperties": false,
        "properties": {
          "decision": {
            "enum": [
              "approve",
              "reject"
            ],
            "title": "Decision",
            "type": "string"
          },
          "note": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Note",
            "type": "string"
          }
        },
        "required": [
          "decision",
          "note"
        ],
        "title": "SkillDraftReviewRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "ToxAgent control plane.\n\nEvidence and decision support over the ToxPred predictor. Every number in an\naccepted answer resolves to a stored observation and a field path; hERG, Tox21\nand ClinTox stay three separate measurements; and no schema here can carry an\naggregate toxicity or safety verdict.\n",
    "title": "ToxAgent control plane",
    "version": "0.1.0.dev0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/skill-drafts": {
      "get": {
        "operationId": "list_skill_drafts_v1_skill_drafts_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 16,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillDraftList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Skill Drafts",
        "tags": [
          "toxagent"
        ]
      },
      "post": {
        "description": "Propose a skill package for expert review. It is validated exactly as a\nshipped skill would be, and never offered to a run.",
        "operationId": "propose_skill_draft_v1_skill_drafts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkillDraftRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillDraft"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Propose Skill Draft",
        "tags": [
          "toxagent"
        ]
      }
    },
    "/v1/skill-drafts/{draft_id}": {
      "get": {
        "operationId": "get_skill_draft_v1_skill_drafts__draft_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "title": "Draft Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillDraft"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Skill Draft",
        "tags": [
          "toxagent"
        ]
      }
    },
    "/v1/skill-drafts/{draft_id}/package": {
      "get": {
        "description": "An approved draft as files, with the manifest made active — the input to\n``scripts/promote_skill_draft.py`` and a reviewed change to the catalog.",
        "operationId": "export_skill_draft_v1_skill_drafts__draft_id__package_get",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "title": "Draft Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillDraftPackage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Skill Draft",
        "tags": [
          "toxagent"
        ]
      }
    },
    "/v1/skill-drafts/{draft_id}:review": {
      "post": {
        "description": "Approve or reject a proposed draft. Needs the expert role; an author\ndoes not review their own draft. Approval does not reach the catalog.",
        "operationId": "review_skill_draft_v1_skill_drafts__draft_id__review_post",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "title": "Draft Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkillDraftReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillDraft"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Review Skill Draft",
        "tags": [
          "toxagent"
        ]
      }
    },
    "/v1/skill-drafts/{draft_id}:withdraw": {
      "post": {
        "operationId": "withdraw_skill_draft_v1_skill_drafts__draft_id__withdraw_post",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "title": "Draft Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillDraft"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Withdraw Skill Draft",
        "tags": [
          "toxagent"
        ]
      }
    },
...
  }
}